+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Add upgrade-deps GH Actions workflow
+- Add --version flag to show version info
+- Add support for Python 3.11 and 3.13
 
 ### Changed
 - Upgrade deps to latest versions